EVHC_gemini_local/requirements.txt
vertexai
uvicorn
fastapi
python-multipart
python-dotenv
